fix: 越城协议修复BMS生产时间上报不对问题、故障码转换增加防雷器、枪座、断路器、水浸、烟感、倾倒、液冷、熔断器

+++ net/net_ycp/src/ycp_fault_analyse.c
@@ -96,7 +96,7 @@ static int32_t ycp_get_fault_code(uint8_t bit, uint8_t gunno)
         return -0x01;
     }
 
-    extern uint8_t ycp_chargepile_fault_converted(uint8_t bit);
+    extern uint16_t ycp_chargepile_fault_converted(uint16_t bit);
     bit = ycp_chargepile_fault_converted(bit);
 
     switch(bit){
